Ironing out LVL 2 mechanics
Changed the pattern for the redScanner and made it to where jumping or
not jumping is necessary to pass instead of movement alone.

Delegating roles in scanner scripts more so that it is more modular.

Both behave as expected but do not destroy by default after their
hashtime is up.

Beginning to separate scanner movement script into its own class.